requirements.txt
absl-py==0.2.0
appnope==0.1.0
astor==0.6.2
awscli==1.14.64
backcall==0.1.0
bleach==1.5.0
botocore==1.9.17
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.7
coremltools==2.1.0
cycler==0.10.0
Cython==0.28.2
decorator==4.3.0
docutils==0.14
easydict==1.7
entrypoints==0.2.3
gast==0.2.0
graphviz==0.8.3
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
imgaug==0.2.5
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2>=2.10.1
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.2
keras-contrib==2.0.8
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
lxml==4.2.3
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mxnet==1.1.0.post0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook>=5.7.2
numpy==1.14.5
opencv-python==3.4.1.15
pandas==0.22.0
pandocfilters==1.4.2
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.2.0
prettytable==0.7.2
prometheus-client==0.3.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pyasn1==0.4.2
pycocotools==2.0.0
pydot==1.2.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.4
PyWavelets==0.5.2
PyYAML>=4.2b1
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.20.0
rsa==3.4.2
s3transfer==0.1.13
scikit-image==0.14.0
scipy==1.1.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.9.0
tensorflow==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tfcoreml==0.2.0
torch==0.4.1
tornado==5.0.2
tqdm==4.23.4
traitlets==4.3.2
turicreate==4.3.2
urllib3>=1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1